feat: extract styles for reusable speakerCard component
