+++ app/components/featuredSpeakers/FeaturedSpeakers.tsx
@@ -5,7 +5,7 @@ import styles from "./featured-speakers.module.css";
 import Image from "next/image";
 import { featuredSpeakers } from "../../data";
 import DetailsBtn from "../detailsBtn/DetailsBtn";
-import SpeakerCard from "../card/SpeakerCard";
+import SpeakerCard from "../speakerCard/SpeakerCard";
 
 const FeaturedSpeakers = () => {
   return (
