+++ app/components/speakerCard/SpeakerCard.tsx
